Split prefix expansion from the rest of the search algorithm
